+++ frontend/src/components/Changelog.vue
@@ -4,6 +4,33 @@
 
     <v-expansion-panels accordion multiple v-model="panel">
       <v-expansion-panel>
+        <v-expansion-panel-header>2.8.0 &mdash; 2023-02-19</v-expansion-panel-header>
+        <v-expansion-panel-content>
+          <v-card flat>
+            <v-card-subtitle>Added</v-card-subtitle>
+            <v-card-text>
+              <ul>
+                  <li>Missiles page</li>
+              </ul>
+            </v-card-text>
+            <v-card-subtitle>Changed</v-card-subtitle>
+            <v-card-text>
+              <ul>
+                  <li>Missiles are no longer displayed in Items page</li>
+              </ul>
+            </v-card-text>
+            <v-card-subtitle>Fixed</v-card-subtitle>
+            <v-card-text>
+              <ul>
+                  <li>Fix DPS calculation for EMP, Fire in Crafts page</li>
+                  <li>Fix Speed calculation in Crafts page</li>
+              </ul>
+            </v-card-text>
+        </v-card>
+        </v-expansion-panel-content>
+      </v-expansion-panel>
+
+      <v-expansion-panel>
         <v-expansion-panel-header>2.7.0 &mdash; 2023-02-17</v-expansion-panel-header>
         <v-expansion-panel-content>
           <v-card flat>