+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.8.0] - 2023-02-19
+
+### Added
+
+- Missiles page
+
+### Changed
+
+- Missiles are no longer displayed in Items page
+
+### Fixed
+
+- Fix DPS calculation for EMP, Fire in Crafts page
+- Fix Speed calculation in Crafts page
+
 ## [2.7.0] - 2023-02-17
 
 ### Added
@@ -292,7 +307,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Last Jeff commit
 
 [unreleased]: https://github.com/solevis/pixyship/compare/main...develop
-[rolling]: https://github.com/solevis/pixyship/compare/2.7.0...main
+[rolling]: https://github.com/solevis/pixyship/compare/2.8.0...main
+[2.8.0]: https://github.com/solevis/pixyship/compare/2.7.0...2.8.0
 [2.7.0]: https://github.com/solevis/pixyship/compare/2.6.0...2.7.0
 [2.6.0]: https://github.com/solevis/pixyship/compare/2.5.0...2.6.0
 [2.5.0]: https://github.com/solevis/pixyship/compare/2.4.0...2.5.0